+++ CHANGELOG.md
@@ -432,3 +432,4 @@
 - 2024-06-14: Test  outreach base, summaries, leads, contacts, +4
 - 2024-06-14: Impl mailbox, email opener, mutate queries, embeddings, +2 query
 - 2024-06-15: Pagination contacts, leads, user, send, +4
+- 2024-06-16: Add mod, user, summaries, leads, +3 cfg